Added the reset password logic.
I actually implemented two logics. The first generates a new password,
then updates the attribute in the user's record and sends an email to
the user with the new password.

The second generates a link that expires after five (5) minutes and
sends this link to the user's registered email. If the user follows
the link, then the system would redirect the user to the 'change
password' view. Because there is no such a view, it only sends a 200
response back saying that it would redirect to the 'change password'
view.